third_party/blink/web_tests/virtual/disable-ua-client-hint/external/wpt/client-hints/accept-ch-stickiness/cross-origin-iframe-redirect-with-fp-delegation.https-expected.txt

This is a testharness.js-based test.
FAIL Iframe redirect with Feature Policy delegation got client hints according to expectations. assert_equals: message from opened frame expected "PASS" but got "UA"
Harness: the test ran to completion.
